requirements/base.txt
sanic==19.12.2
python-decouple==3.3
SQLAlchemy==1.3.12
alembic==1.3.2
aiotask-context==0.6.0
Cerberus==1.3.2
